fix alias analysis, correctly propagate escapes via nested allocations
